refactor exception handling & closing I/O resources

+++ flying-saucer-core/src/main/java/org/xhtmlrenderer/util/IOUtil.java
@@ -1,56 +1,56 @@
 package org.xhtmlrenderer.util;
 
-import java.io.*;
+import javax.annotation.CheckReturnValue;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
+import java.io.BufferedInputStream;
+import java.io.BufferedOutputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.Closeable;
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.net.URL;
 import java.net.URLConnection;
+import java.util.logging.Level;
+
+import static java.nio.file.Files.newInputStream;
+import static java.nio.file.Files.newOutputStream;
 
 /**
  * @author patrick
  */
+@ParametersAreNonnullByDefault
 public class IOUtil {
-    public static File copyFile(File page, File outputDir) throws IOException {
-        InputStream in = null;
-        OutputStream out = null;
-        try {
-            in = new BufferedInputStream(new FileInputStream(page));
-            File outputFile;
-            outputFile = new File(outputDir, page.getName());
-            out = new BufferedOutputStream(new FileOutputStream(outputFile));
+    public static void copyFile(File page, File outputDir) throws IOException {
+        File outputFile = new File(outputDir, page.getName());
 
-            // Transfer bytes from in to out
-            byte[] buf = new byte[1024];
-            int len;
-            while ((len = in.read(buf)) > 0) {
-                out.write(buf, 0, len);
+        try (OutputStream out = new BufferedOutputStream(newOutputStream(outputFile.toPath()))) {
+            try (InputStream in = new BufferedInputStream(newInputStream(page.toPath()))) {
+                copyBytes(in, out);
             }
-            in.close();
             out.flush();
-            out.close();
-            return outputFile;
-        } finally {
-            if (in != null) {
-                try {
-                    in.close();
-                } catch (IOException e) {
-                    // swallow
-                }
-            }
-            if (out != null) {
-                try {
-                    out.close();
-                } catch (IOException e) {
-                    // swallow
-                }
-            }
+        }
+    }
+
+    private static void copyBytes(InputStream in, OutputStream out) throws IOException {
+        byte[] buf = new byte[1024];
+        int len;
+        while ((len = in.read(buf)) >= 0) {
+            out.write(buf, 0, len);
         }
     }
 
     public static void deleteAllFiles(final File dir) throws IOException {
         File[] files = dir.listFiles();
-        for (int i = 0; i < files.length; i++) {
-            File file = files[i];
-            if (!file.delete()) {
-                throw new IOException("Cleanup directory " + dir + ", can't delete file " + file);
+        if (files != null) {
+            for (File file : files) {
+                if (!file.delete()) {
+                    throw new IOException("Cleanup directory " + dir + ", can't delete file " + file);
+                }
             }
         }
     }
@@ -61,7 +61,6 @@ public class IOUtil {
      * buffer the stream.
      */
     public static InputStream openStreamAtUrl(String uri) {
-        InputStream is = null;
         try {
             final URLConnection uc = new URL(uri).openConnection();
 
@@ -82,7 +81,7 @@ public class IOUtil {
 
             uc.connect();
 
-            is = uc.getInputStream();
+            return uc.getInputStream();
         } catch (java.net.MalformedURLException e) {
             XRLog.exception("bad URL given: " + uri, e);
         } catch (FileNotFoundException e) {
@@ -91,7 +90,55 @@ public class IOUtil {
             XRLog.exception("IO problem for " + uri, e);
         }
 
-        return is;
+        return null;
+    }
 
+    /**
+     * Gets a Reader for the resource identified
+     *
+     * @return The stylesheet value
+     */
+    @Nullable
+    @CheckReturnValue
+    public static InputStream getInputStream(String uri) {
+        try {
+            return new URL(uri).openStream();
+        } catch (java.net.MalformedURLException e) {
+            XRLog.exception("bad URL given: " + uri, e);
+        } catch (java.io.FileNotFoundException e) {
+            XRLog.exception("item at URI " + uri + " not found");
+        } catch (java.io.IOException e) {
+            XRLog.exception("IO problem for " + uri, e);
+        }
+        return null;
+    }
+
+    @Nullable
+    @CheckReturnValue
+    public static byte[] readBytes(String uri) {
+        try (InputStream is = getInputStream(uri)) {
+            if (is == null) return null;
+            return readBytes(is);
+        } catch (IOException e) {
+            XRLog.load(Level.WARNING, "Unable to read " + uri, e);
+            return null;
+        }
+    }
+
+    @Nonnull
+    @CheckReturnValue
+    public static byte[] readBytes(InputStream is) throws IOException {
+        ByteArrayOutputStream result = new ByteArrayOutputStream();
+        copyBytes(is, result);
+        return result.toByteArray();
+    }
+
+    public static void close(@Nullable Closeable in) {
+        if (in != null) {
+            try {
+                in.close();
+            } catch (IOException ignore) {
+            }
+        }
     }
 }
